move tests to separate dialogbase file and add more error message validation
