+++ extensions/data-workspace/src/test/dialogs/newProjectDialog.test.ts
@@ -9,7 +9,6 @@ import * as os from 'os';
 import * as path from 'path';
 import * as vscode from 'vscode';
 import * as sinon from 'sinon';
-import * as utils from '../../common/utils';
 import { promises as fs } from 'fs';
 import { NewProjectDialog } from '../../dialogs/newProjectDialog';
 import { WorkspaceService } from '../../services/workspaceService';
@@ -43,38 +42,6 @@ suite('New Project Dialog', function (): void {
 		should.equal(await dialog.validate(), true, 'Validation should pass because name is unique and parent directory exists');
 	});
 
-	test('Should validate new workspace location', async function (): Promise<void> {
-		const workspaceServiceMock = TypeMoq.Mock.ofType<WorkspaceService>();
-		workspaceServiceMock.setup(x => x.getAllProjectTypes()).returns(() => Promise.resolve([testProjectType]));
-
-		const dialog = new NewProjectDialog(workspaceServiceMock.object);
-		await dialog.open();
-
-		dialog.model.name = `TestProject_${new Date().getTime()}`;
-		dialog.model.location = os.tmpdir();
-		dialog.workspaceInputBox!.value = 'test';
-		should.equal(await dialog.validate(), false, 'Validation should fail because workspace does not end in .code-workspace');
-
-		const fileExistStub = sinon.stub(utils, 'fileExist');
-		fileExistStub.resolves(true);
-		// use invalid folder
-		dialog.workspaceInputBox!.value = 'invalidLocation/test.code-workspace';
-		should.equal(await dialog.validate(), false, 'Validation should fail because the folder is invalid');
-
-		// use already existing workspace
-		const existingWorkspaceFilePath = path.join(os.tmpdir(), `${dialog.model.name}.code-workspace`);
-		dialog.workspaceInputBox!.value = existingWorkspaceFilePath;
-		should.equal(await dialog.validate(), false, 'Validation should fail because the selected workspace file already exists');
-
-		// same folder as the project should be valid even if the project folder isn't created yet
-		fileExistStub.resolves(false);
-		dialog.workspaceInputBox!.value = path.join(dialog.model.location, dialog.model.name, 'test.code-workspace');
-		should.equal(await dialog.validate(), true, `Validation should pass if the file location is the same folder as the project. Error was: ${dialog.getErrorMessage().text}`);
-
-		// change workspace name to something that should pass
-		dialog.workspaceInputBox!.value = path.join(os.tmpdir(), `TestWorkspace_${new Date().getTime()}.code-workspace`);
-		should.equal(await dialog.validate(), true, `Validation should pass because the parent directory exists, workspace filepath is unique, and the file extension is correct. Error was: ${dialog.getErrorMessage().text}`);
-	});
 
 	test('Should validate workspace in onComplete', async function (): Promise<void> {
 		const workspaceServiceMock = TypeMoq.Mock.ofType<WorkspaceService>();
